Completed Section 23

+++ DJANGO_COURSE_2.xx/20-Blog_Clone_Project/mysite/blog/views.py
@@ -1,15 +1,18 @@
 from django.shortcuts import render, get_object_or_404, redirect
 from django.contrib.auth.decorators import login_required
 from django.utils import timezone
+from django.http.response import HttpResponseNotAllowed, HttpResponseRedirect, HttpResponseForbidden
+
 from .models import Post, Comment
 from .forms import PostForm, CommentForm
 
-from django.views.generic import (TemplateView,ListView,
-                                  DetailView,CreateView,
-                                  UpdateView,DeleteView)
+from django.views.generic import (
+    TemplateView, ListView, DetailView, CreateView, UpdateView, DeleteView
+)
 
 from django.urls import reverse_lazy
 from django.contrib.auth.mixins import LoginRequiredMixin
+
 
 # Create your views here.
 class AboutView(TemplateView):
@@ -22,11 +25,26 @@ class PostListView(ListView):
     def get_queryset(self):
         return Post.objects.filter(published_date__lte=timezone.now()).order_by('-published_date')
 
+
 class PostDetailView(DetailView):
     model = Post
 
 
-class CreatePostView(LoginRequiredMixin,CreateView):
+class CreatePostView(LoginRequiredMixin, CreateView):
+    login_url = '/login/'
+    redirect_field_name = 'blog/post_detail.html'
+
+    form_class = PostForm
+
+    model = Post
+
+    # Set author as default value to request.user
+    def form_valid(self, form):
+        form.instance.author = self.request.user
+        return super(CreatePostView, self).form_valid(form)
+
+
+class PostUpdateView(LoginRequiredMixin, UpdateView):
     login_url = '/login/'
     redirect_field_name = 'blog/post_detail.html'
 
@@ -35,18 +53,18 @@ class CreatePostView(LoginRequiredMixin,CreateView):
     model = Post
 
 
-class PostUpdateView(LoginRequiredMixin,UpdateView):
-    login_url = '/login/'
-    redirect_field_name = 'blog/post_detail.html'
-
-    form_class = PostForm
-
+class PostDeleteView(LoginRequiredMixin, DeleteView):
     model = Post
 
-
-class PostDeleteView(LoginRequiredMixin,DeleteView):
-    model = Post
-    success_url = reverse_lazy('post_list')
+    def delete(self, request, *args, **kwargs):
+        # a post can only be removed by author
+        post = self.get_object()
+        if post.author == request.user:
+            success_url = reverse_lazy('post_list')
+            post.delete()
+            return HttpResponseRedirect(success_url)
+        else:
+            return HttpResponseForbidden("Cannot delete other's posts")
 
 
 class DraftListView(LoginRequiredMixin,ListView):
@@ -56,18 +74,20 @@ class DraftListView(LoginRequiredMixin,ListView):
     model = Post
 
     def get_queryset(self):
+        # posts without published_date
         return Post.objects.filter(published_date__isnull=True).order_by('created_date')
 
 
-#######################################
-## Functions that require a pk match ##
-#######################################
+#########################################
+#   Functions that require a pk match   #
+#########################################
 
 @login_required
 def post_publish(request, pk):
     post = get_object_or_404(Post, pk=pk)
     post.publish()
     return redirect('post_detail', pk=pk)
+
 
 @login_required
 def add_comment_to_post(request, pk):
@@ -77,6 +97,7 @@ def add_comment_to_post(request, pk):
         if form.is_valid():
             comment = form.save(commit=False)
             comment.post = post
+            comment.author = request.user
             comment.save()
             return redirect('post_detail', pk=post.pk)
     else:
@@ -94,6 +115,8 @@ def comment_approve(request, pk):
 @login_required
 def comment_remove(request, pk):
     comment = get_object_or_404(Comment, pk=pk)
+    if comment.author != request.user:
+        return HttpResponseNotAllowed("You are only allowed to remove your own comments!")
     post_pk = comment.post.pk
     comment.delete()
     return redirect('post_detail', pk=post_pk)
